+++ Configuration/Eras/python/Modifier_ctpps_2022_cff.py
@@ -1,0 +1,3 @@
+import FWCore.ParameterSet.Config as cms
+
+ctpps_2022 =  cms.Modifier()
